+++ doomtype.h
@@ -37,6 +37,8 @@ typedef unsigned char byte;
 #endif
 
 
+#include <stdint.h>
+
 // Predefined with some OS.
 #ifdef LINUX
 #include <values.h>
@@ -46,13 +48,13 @@ typedef unsigned char byte;
 
 // Max pos 32-bit int.
 #define MAXINT		((int)0x7fffffff)	
-#define MAXLONG		((long)0x7fffffff)
+#define MAXLONG		((int32_t)0x7fffffff)
 #define MINCHAR		((char)0x80)
 #define MINSHORT	((short)0x8000)
 
 // Max negative 32-bit integer.
 #define MININT		((int)0x80000000)	
-#define MINLONG		((long)0x80000000)
+#define MINLONG		((int32_t)0x80000000)
 #endif
